Split screen.rs into dir structure
